pyproject.toml
[project]
name = "mcp-weather"
version = "0.1.0"
description = "An mcp server for weather information."
authors = [
    {name = "Melvin Klein",email = "70433111+MelvinKl@users.noreply.github.com"}
]
license = {text = "Apache2"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "mcp (>=1.9.1,<2.0.0)",
    "weatherapipython @ git+https://github.com/weatherapicom/python.git"
]


[tool.flake8]
exclude= [".eggs", ".git", ".hg", ".mypy_cache", ".tox", ".venv", ".devcontainer", "venv", "_build", "buck-out", "build", "dist", "**/__init__.py"]
statistics = true
show-source = false
max-complexity = 8
max-annotations-complexity = 3
docstring-convention = 'numpy'
max-line-length = 120
ignore = ["E203", "W503", "E704"]
inline-quotes = '"'
docstring-quotes = '"""'
multiline-quotes = '"""'
dictionaries = ["en_US", "python", "technical", "pandas"]
ban-relative-imports = true
per-file-ignores = """
"""

[tool.black]
line-length = 120
exclude = """
/(
      .eggs
    | .git
    | .hg
    | .mypy_cache
    | .nox
    | .pants.d
    | .tox
    | .venv
    | _build
    | buck-out
    | build
    | dist
    | node_modules
    | venv
)/
"""

[tool.isort]
profile = "black"
skip = ['.eggs', '.git', '.hg', '.mypy_cache', '.nox', '.pants.d', '.tox', '.venv', '_build', 'buck-out', 'build', 'dist', 'node_modules', 'venv']
skip_gitignore = true
known_local_folder = []

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.14"
pytest = "^8.3.5"
coverage = "^7.8.0"
flake8 = "^7.2.0"
flake8-black = "^0.3.6"
flake8-pyproject = "^1.2.3"
flake8-quotes = "^3.4.0"
flake8-return = "^1.2.0"
flake8-annotations-complexity = "^0.1.0"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^24.12.12"
flake8-builtins = "^2.5.0"
flake8-comprehensions = "^3.15.0"
flake8-eradicate = "^1.5.0"
flake8-expression-complexity = "^0.0.11"
flake8-pytest-style = "^2.1.0"
pep8-naming = "^0.15.1"
flake8-eol = "^0.0.8"
flake8-exceptions = "^0.0.1a0"
flake8-simplify = "^0.21.0"
flake8-wot = "^0.2.0"
flake8-function-order = "^0.0.5"
flake8-tidy-imports = "^4.10.0"
black = "^25.1.0"